add menu, changed logic of Game

+++ NumberGeneration.cpp
@@ -9,7 +9,7 @@ std::random_device NumberGeneration::rand;
 std::mt19937 NumberGeneration::gen(NumberGeneration::rand());
 
 size_t NumberGeneration::generateRandomNumber(size_t min, size_t max) {
-    std::uniform_int_distribution<> dis(min, max);
+    std::uniform_int_distribution<> dis(static_cast<int>(min), static_cast<int>(max));
     return dis(gen);
 }
 
@@ -26,6 +26,6 @@ size_t NumberGeneration::generateRandomOddNumber(size_t min, size_t max) {
         std::swap(min, max);
     }
 
-    std::uniform_int_distribution<> dis(0, (max - min) / 2);
+    std::uniform_int_distribution<> dis(0, static_cast<int>(max - min) / 2);
     return dis(gen) * 2 + min;
 }
